+++ waspc/data/packages/wasp-config/src/spec/publicApi/constructors.ts
@@ -1,0 +1,35 @@
+import type { Action, App, Page, Query, Route } from "./tsAppSpec.js";
+
+export function app(input: Omit<App, "kind">): App {
+  return input;
+}
+
+export function page(
+  component: Page["component"],
+  config?: Pick<Page, "authRequired">,
+): Page {
+  return { kind: "page", component, ...config };
+}
+
+export function route(
+  name: Route["name"],
+  path: Route["path"],
+  page: Route["page"],
+  config?: Pick<Route, "lazy" | "prerender">,
+): Route {
+  return { kind: "route", name, path, page, ...config };
+}
+
+export function query(
+  fn: Query["fn"],
+  config?: Pick<Query, "entities" | "auth">,
+): Query {
+  return { kind: "query", fn, ...config };
+}
+
+export function action(
+  fn: Action["fn"],
+  config?: Pick<Action, "entities" | "auth">,
+): Action {
+  return { kind: "action", fn, ...config };
+}