+++ waspc/data/packages/wasp-config/src/spec/publicApi/index.ts
@@ -1,0 +1,10 @@
+export { action, app, page, query, route } from "./constructors.js";
+export type {
+  Action,
+  App,
+  ExtImport,
+  Page,
+  Part,
+  Query,
+  Route,
+} from "./tsAppSpec.js";