+++ waspc/e2e-tests/Tests/ViteConfigTest.hs
@@ -1,5 +1,3 @@
-{-# LANGUAGE OverloadedRecordDot #-}
-
 module Tests.ViteConfigTest (viteConfigTest) where
 
 import Control.Monad.Reader (ask)
@@ -7,15 +5,15 @@ import NeatInterpolation (trimming)
 import ShellCommands
   ( ShellCommand,
     ShellCommandBuilder,
-    WaspNewTemplate (..),
     WaspProjectContext (..),
-    createFile,
     createTestWaspProject,
     inTestWaspProjectDir,
     waspCliCompile,
+    writeToFile,
   )
 import StrongPath (relfile, (</>))
 import Test (Test (..), TestCase (..))
+import Wasp.Cli.Command.CreateNewProject.AvailableTemplates (minimalStarterTemplate)
 
 viteConfigTest :: Test
 viteConfigTest =
@@ -24,7 +22,7 @@ viteConfigTest =
     [ TestCase
         "fail-on-missing-vite-config"
         ( sequence
-            [ createTestWaspProject Minimal,
+            [ createTestWaspProject minimalStarterTemplate,
               inTestWaspProjectDir
                 [ deleteViteConfig,
                   expectCommandFailure <$> waspCliCompile
@@ -34,7 +32,7 @@ viteConfigTest =
       TestCase
         "fail-on-missing-wasp-plugin-import"
         ( sequence
-            [ createTestWaspProject Minimal,
+            [ createTestWaspProject minimalStarterTemplate,
               inTestWaspProjectDir
                 [ writeViteConfigWithoutPlugin,
                   expectCommandFailure <$> waspCliCompile
@@ -49,7 +47,7 @@ deleteViteConfig = return "rm vite.config.ts"
 writeViteConfigWithoutPlugin :: ShellCommandBuilder WaspProjectContext ShellCommand
 writeViteConfigWithoutPlugin = do
   context <- ask
-  createFile
+  writeToFile
     (context.waspProjectDir </> [relfile|vite.config.ts|])
     [trimming|
       import { defineConfig } from "vite";