+++ waspc/src/Wasp/Generator/SdkGenerator/Server/JobGenerator.hs
@@ -112,7 +112,8 @@ getImportJsonForJobDefinition jobName =
   GJI.jsImportToImportJson $
     Just $
       JI.JsImport
-        { JI._path = JI.ModuleImportPath $ makeSdkImportPath [relfileP|server/jobs|],
+        { JI._kind = JI.ValueImport,
+          JI._path = JI.ModuleImportPath $ makeSdkImportPath [relfileP|server/jobs|],
           JI._name = JI.JsImportField jobName,
           -- NOTE: We are using alias to avoid name conflicts with user defined imports.
           JI._importAlias = Just "_waspJobDefinition"